+++ src/relstorage/adapters/adapter.py
@@ -38,7 +38,7 @@ from ..options import Options
 from ._util import DatabaseHelpersMixin
 from .drivers import _select_driver
 from .interfaces import UnableToLockRowsToModifyError
-from .interfaces import UnableToLockRowsToReadCurrentError
+from .interfaces import UnableToLockRowsDeadlockError
 
 logger = __import__('logging').getLogger(__name__)
 
@@ -192,17 +192,17 @@ class AbstractAdapter(DatabaseHelpersMixin):
             # various concurrency situations.
             return self._composed_lock_objects_and_detect_conflicts(cursor,
                                                                     read_current_oids)
-        begin = time.time()
         try:
             return self._best_lock_objects_and_detect_conflicts(cursor, read_current_oids)
-        except self.locker.lock_exceptions:
-            # Heuristic to guess. If the stored proc or stored proc runner can do better,
-            # they should.
-            elapsed = time.time() - begin
+        except self.locker.lock_exceptions as ex:
+            # Heuristic to guess.
+            # XXX: we should do a lot better. We used to be time based, but since we take
+            # exclusive locks first, that's now useless.
             kind = UnableToLockRowsToModifyError
-            if read_current_oids and elapsed < self.locker.commit_lock_timeout:
-                kind = UnableToLockRowsToReadCurrentError
+            if self.driver.exception_is_deadlock(ex):
+                kind = UnableToLockRowsDeadlockError
 
+            del ex
             self.locker.reraise_commit_lock_error(
                 cursor,
                 self._describe_best_lock_objects_and_detect_conflicts(),
@@ -217,6 +217,7 @@ class AbstractAdapter(DatabaseHelpersMixin):
             # We go ahead and compare the readCurrent TIDs here, so
             # that we don't have to make the call to detect conflicts
             # or even lock rows if there are readCurrent violations.
+            # Recall this function is called twice.
             for oid_int, expect_tid_int in read_current_oids.items():
                 actual_tid_int = current.get(oid_int, 0)
                 if actual_tid_int != expect_tid_int:
@@ -229,7 +230,6 @@ class AbstractAdapter(DatabaseHelpersMixin):
             cursor, read_current_oid_ints,
             self.force_lock_readCurrent_for_share_blocking,
             after_lock_share)
-
         conflicts = self.mover.detect_conflict(cursor)
         return conflicts